move cell and controller to subpaclage
